backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
celery==5.4.0
redis==5.0.8
psycopg2-binary==2.9.9
msgraph-sdk==1.5.4
azure-identity==1.17.1
msal==1.28.0
openai==1.35.3
anthropic==0.30.0
sendgrid==6.11.0
sentry-sdk==2.8.0
gunicorn==22.0.0
whitenoise==6.7.0
cryptography==42.0.8
python-dateutil==2.9.0
